Refactor program graph module for better code organization
